requirements.txt
python-telegram-bot==21.9
pillow